% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getsummary.R
\name{getsummary}
\alias{getsummary}
\title{Summarize an article}
\usage{
getsummary(text)
}
\arguments{
\item{text}{character string to be summarized}
}
\value{
two summaries of the text
}
\description{
Summarize an article
}

